+++ packages/components/__tests__/Link.spec.tsx
@@ -30,4 +30,18 @@ describe('Link', () => {
 		expect(screen.getByRole('link')).toBeVisible();
 		expect(screen.getByRole('link')).toHaveAttribute('href', 'https://www.test.com');
 	});
+
+	it('renders with underline="always"', () => {
+		const navigate = vi.fn();
+		render(
+			<RouterProvider navigate={navigate} useHref={useHref}>
+				<MemoryRouter>
+					<Link href="/test" underline="always">
+						Link
+					</Link>
+				</MemoryRouter>
+			</RouterProvider>,
+		);
+		expect(screen.getByRole('link')).toBeVisible();
+	});
 });