+++ packages/components/src/index.ts
@@ -32,6 +32,7 @@ export type { DisclosurePanelProps, DisclosureProps } from './Disclosure';
 export type { DisclosureGroupProps } from './DisclosureGroup';
 export type { DropIndicatorProps } from './DropIndicator';
 export type { DropZoneProps } from './DropZone';
+export type { EmptyStateProps } from './EmptyState';
 export type { FieldErrorProps } from './FieldError';
 export type { FieldGroupProps } from './FieldGroup';
 export type { FileTriggerProps } from './FileTrigger';
@@ -83,7 +84,7 @@ export type { TextAreaProps } from './TextArea';
 export type { TextFieldProps } from './TextField';
 export type { ToastOptions, ToastValue } from './Toast';
 export type { ToggleButtonProps } from './ToggleButton';
-export type { ToggleButtonGroupProps } from './ToggleButtonGroup';
+export type { ToggleButtonGroupProps, ToggleButtonGroupVariants } from './ToggleButtonGroup';
 export type { ToggleIconButtonProps } from './ToggleIconButton';
 export type { ToolbarProps } from './Toolbar';
 export type { TooltipProps, TooltipTriggerProps } from './Tooltip';
@@ -176,6 +177,7 @@ export {
 export { DisclosureGroup, disclosureGroupStyles } from './DisclosureGroup';
 export { DropIndicator, dropIndicatorStyles } from './DropIndicator';
 export { DropZone, DropZoneContext, dropZoneStyles } from './DropZone';
+export { EmptyState, emptyStateStyles } from './EmptyState';
 export { FieldError, FieldErrorContext, fieldErrorStyles } from './FieldError';
 export { FieldGroup, fieldGroupStyles } from './FieldGroup';
 export { FileTrigger } from './FileTrigger';
@@ -319,7 +321,7 @@ export {
 	toastRegionStyles,
 	toastStyles,
 } from './Toast';
-export { ToggleButton, ToggleButtonContext } from './ToggleButton';
+export { ToggleButton, ToggleButtonContext, toggleButtonElevatedStyles } from './ToggleButton';
 export {
 	ToggleButtonGroup,
 	ToggleButtonGroupContext,